requirements.txt
django
django-ninja
easyocr
pillow
